# apps/api/openapi/openapi.json
{
  "components": {
    "schemas": {
      "CsrfTokenDataDto": {
        "properties": {
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "token": {
            "description": "Non-secret CSRF token bound to the anonymous browser transaction.",
            "type": "string"
          }
        },
        "required": [
          "token",
          "expiresAt"
        ],
        "type": "object"
      },
      "CsrfTokenResponseDto": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CsrfTokenDataDto"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
      "EmailVerificationRequestAcceptedDataDto": {
        "properties": {
          "status": {
            "enum": [
              "VERIFICATION_EMAIL_SENT_IF_ELIGIBLE"
            ],
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "type": "object"
      },
      "EmailVerificationRequestAcceptedResponseDto": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/EmailVerificationRequestAcceptedDataDto"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
      "EmailVerificationRequestDto": {
        "properties": {
          "email": {
            "example": "kullanici@example.com",
            "format": "email",
            "maxLength": 254,
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "type": "object"
      },
      "RegisterAccountRequestDto": {
        "properties": {
          "email": {
            "example": "kullanici@example.com",
            "format": "email",
            "maxLength": 254,
            "type": "string"
          },
          "password": {
            "format": "password",
            "maxLength": 128,
            "minLength": 12,
            "type": "string",
            "writeOnly": true
          },
          "passwordConfirmation": {
            "format": "password",
            "maxLength": 128,
            "minLength": 12,
            "type": "string",
            "writeOnly": true
          },
          "termsAccepted": {
            "enum": [
              true
            ],
            "type": "boolean"
          }
        },
        "required": [
          "email",
          "password",
          "passwordConfirmation",
          "termsAccepted"
        ],
        "type": "object"
      },
      "RegistrationAcceptedDataDto": {
        "properties": {
          "next": {
            "example": "/verify-email",
            "type": "string"
          },
          "status": {
            "enum": [
              "VERIFICATION_REQUIRED"
            ],
            "type": "string"
          }
        },
        "required": [
          "status",
          "next"
        ],
        "type": "object"
      },
      "RegistrationAcceptedResponseDto": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RegistrationAcceptedDataDto"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
      "VerifyEmailDataDto": {
        "properties": {
          "next": {
            "example": "/login",
            "type": "string"
          },
          "status": {
            "enum": [
              "VERIFIED"
            ],
            "type": "string"
          }
        },
        "required": [
          "status",
          "next"
        ],
        "type": "object"
      },
      "VerifyEmailRequestDto": {
        "properties": {
          "code": {
            "example": "12345678",
            "maxLength": 8,
            "minLength": 8,
            "pattern": "^\\d{8}$",
            "type": "string",
            "writeOnly": true
          },
          "email": {
            "example": "kullanici@example.com",
            "format": "email",
            "maxLength": 254,
            "type": "string"
          }
        },
        "required": [
          "email",
          "code"
        ],
        "type": "object"
      },
      "VerifyEmailResponseDto": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/VerifyEmailDataDto"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
      "VersionResponseDto": {
        "properties": {
          "version": {
            "example": "0.0.0",
            "type": "string"
          }
        },
        "required": [
          "version"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "REST contract for the personal task planner.",
    "title": "Personal Task Planner API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/csrf": {
      "get": {
        "operationId": "getAuthCsrf",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CsrfTokenResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Issue a CSRF token for an anonymous authentication transaction",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/email-verification-requests": {
      "post": {
        "operationId": "requestEmailVerification",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailVerificationRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailVerificationRequestAcceptedResponseDto"
                }
              }
            },
            "description": ""
          },
          "422": {
            "description": "Safe validation details."
          },
          "429": {
            "description": "Generic verification request rate limit."
          }
        },
        "summary": "Request a fresh verification email without account enumeration",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/email-verifications": {
      "post": {
        "operationId": "verifyEmail",
        "parameters": [
          {
            "description": "A unique key for this verification attempt.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyEmailResponseDto"
                }
              }
            },
            "description": ""
          },
          "409": {
            "description": "The code was already used or the idempotent request is still processing."
          },
          "422": {
            "description": "Safe validation or invalid/expired-code details."
          },
          "429": {
            "description": "Generic verification rate limit."
          }
        },
        "summary": "Verify a pending email identity with a manual code",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "operationId": "registerAccount",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterAccountRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationAcceptedResponseDto"
                }
              }
            },
            "description": ""
          },
          "422": {
            "description": "Safe validation details."
          },
          "429": {
            "description": "Generic registration rate limit."
          }
        },
        "summary": "Submit an email/password registration without account enumeration",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/version": {
      "get": {
        "operationId": "getVersion",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Return the running contract version",
        "tags": [
          "system"
        ]
      }
    }
  },
  "security": [],
  "servers": [
    {
      "url": "/"
    }
  ],
  "tags": []
}