# packages/api-client/src/generated/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
    baseUrl: `${string}://${string}` | (string & {});
};

export type CsrfTokenDataDto = {
    expiresAt: string;
    /**
     * Non-secret CSRF token bound to the anonymous browser transaction.
     */
    token: string;
};

export type CsrfTokenResponseDto = {
    data: CsrfTokenDataDto;
};

export type EmailVerificationRequestAcceptedDataDto = {
    status: 'VERIFICATION_EMAIL_SENT_IF_ELIGIBLE';
};

export type EmailVerificationRequestAcceptedResponseDto = {
    data: EmailVerificationRequestAcceptedDataDto;
};

export type EmailVerificationRequestDto = {
    email: string;
};

export type RegisterAccountRequestDto = {
    email: string;
    termsAccepted: true;
};

export type RegistrationAcceptedDataDto = {
    next: string;
    status: 'VERIFICATION_REQUIRED';
};

export type RegistrationAcceptedResponseDto = {
    data: RegistrationAcceptedDataDto;
};

export type VerifyEmailDataDto = {
    next: string;
    status: 'VERIFIED';
};

export type VerifyEmailRequestDto = {
    email: string;
};

export type VerifyEmailResponseDto = {
    data: VerifyEmailDataDto;
};

export type VersionResponseDto = {
    version: string;
};

export type RegisterAccountRequestDtoWritable = {
    email: string;
    password: string;
    passwordConfirmation: string;
    termsAccepted: true;
};

export type VerifyEmailRequestDtoWritable = {
    code: string;
    email: string;
};

export type GetAuthCsrfData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/api/v1/auth/csrf';
};

export type GetAuthCsrfResponses = {
    200: CsrfTokenResponseDto;
};

export type GetAuthCsrfResponse = GetAuthCsrfResponses[keyof GetAuthCsrfResponses];

export type RequestEmailVerificationData = {
    body: EmailVerificationRequestDto;
    headers: {
        'X-CSRF-Token': string;
    };
    path?: never;
    query?: never;
    url: '/api/v1/auth/email-verification-requests';
};

export type RequestEmailVerificationErrors = {
    /**
     * Safe validation details.
     */
    422: unknown;
    /**
     * Generic verification request rate limit.
     */
    429: unknown;
};

export type RequestEmailVerificationResponses = {
    202: EmailVerificationRequestAcceptedResponseDto;
};

export type RequestEmailVerificationResponse = RequestEmailVerificationResponses[keyof RequestEmailVerificationResponses];

export type VerifyEmailData = {
    body: VerifyEmailRequestDtoWritable;
    headers: {
        /**
         * A unique key for this verification attempt.
         */
        'Idempotency-Key': string;
        'X-CSRF-Token': string;
    };
    path?: never;
    query?: never;
    url: '/api/v1/auth/email-verifications';
};

export type VerifyEmailErrors = {
    /**
     * The code was already used or the idempotent request is still processing.
     */
    409: unknown;
    /**
     * Safe validation or invalid/expired-code details.
     */
    422: unknown;
    /**
     * Generic verification rate limit.
     */
    429: unknown;
};

export type VerifyEmailResponses = {
    200: VerifyEmailResponseDto;
};

export type VerifyEmailResponse = VerifyEmailResponses[keyof VerifyEmailResponses];

export type RegisterAccountData = {
    body: RegisterAccountRequestDtoWritable;
    headers: {
        'X-CSRF-Token': string;
    };
    path?: never;
    query?: never;
    url: '/api/v1/auth/register';
};

export type RegisterAccountErrors = {
    /**
     * Safe validation details.
     */
    422: unknown;
    /**
     * Generic registration rate limit.
     */
    429: unknown;
};

export type RegisterAccountResponses = {
    202: RegistrationAcceptedResponseDto;
};

export type RegisterAccountResponse = RegisterAccountResponses[keyof RegisterAccountResponses];

export type GetVersionData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/api/v1/version';
};

export type GetVersionResponses = {
    200: VersionResponseDto;
};

export type GetVersionResponse = GetVersionResponses[keyof GetVersionResponses];